+++ framework/shell/components/app-hub/composables/index.ts
@@ -1,0 +1,1 @@
+export { useAppHub } from "@shell/components/app-hub/composables/useAppHub";