+++ framework/shell/components/app-switcher/composables/index.ts
@@ -1,1 +1,0 @@
-export { useAppSwitcher } from "@shell/components/app-switcher/composables/useAppSwitcher";